+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Rename the 'cache' subcommand to 'workspace' because it wasn't clear that users should actually work inside cache directories (i.e. they're not temporary caches).
 * Colourise output messages for clarity
 * Add a `--no-color` option to the main sugarkube command to disable coloured output
+* The DAG now prints progress information about the nodes it's waiting on.
 
 ## 0.8.0
 * Pinning kapp versions in stacks is now much more concise. See `internal/testdata/stack-pinned.yaml` for an example.